+++ pyroomacoustics/bss/fastmnmf.py
@@ -23,6 +23,7 @@ FastMNMF
 
 Blind Source Separation using Fast Multichannel Nonnegative Matrix Factorization (FastMNMF)
 """
+
 import numpy as np