+++ pyroomacoustics/doa/doa.py
@@ -323,8 +323,7 @@ class DOA(object):
             self.angle_of_arrival = None
         if X.shape[0] != self.M:
             raise ValueError(
-                "Number of signals (rows) does not match the \
-                number of microphones."
+                "Number of signals (rows) does not match the number of microphones."
             )
         if X.shape[1] != self.max_bin:
             raise ValueError("Mismatch in FFT length.")
@@ -577,8 +576,7 @@ class DOA(object):
         #     num_src = self.M
         if num_src < 1:
             warnings.warn(
-                "Number of sources must be at least 1. Changing \
-                number of sources to 1."
+                "Number of sources must be at least 1. Changing number of sources to 1."
             )
             num_src = 1
         valid = int(num_src)