+++ pyroomacoustics/doa/tools_fri_doa_plane.py
@@ -1788,7 +1788,7 @@ def pt_src_recon_multiband(
     if len(a.shape) == 2:
         assert a.shape[1] == num_bands
 
-    a_ri = np.row_stack((a.real, a.imag))
+    a_ri = np.vstack((a.real, a.imag))
 
     if update_G:
         if "G_iter" in kwargs:
@@ -2097,7 +2097,7 @@ def pt_src_recon_rotate(
         # rotate microphone steering vector
         # (due to change of coordinate w.r.t. the random rotation)
         p_mic_xy_rotated = np.dot(
-            rotate_mtx, np.row_stack((p_mic_x.flatten("F"), p_mic_y.flatten("F")))
+            rotate_mtx, np.vstack((p_mic_x.flatten("F"), p_mic_y.flatten("F")))
         )
         p_mic_x_rotated = np.reshape(p_mic_xy_rotated[0, :], p_mic_x.shape, order="F")
         p_mic_y_rotated = np.reshape(p_mic_xy_rotated[1, :], p_mic_y.shape, order="F")
@@ -2137,7 +2137,7 @@ def pt_src_recon_rotate(
                 xk_recon_rotated, yk_recon_rotated = polar2cart(1, phik_recon_rotated)
                 xy_rotate_back = linalg.solve(
                     rotate_mtx,
-                    np.row_stack(
+                    np.vstack(
                         (xk_recon_rotated.flatten("F"), yk_recon_rotated.flatten("F"))
                     ),
                 )