+++ pyroomacoustics/experimental/rt60.py
@@ -30,6 +30,7 @@ References
 .. [1] M. R. Schroeder, "New Method of Measuring Reverberation Time,"
     J. Acoust. Soc. Am., vol. 37, no. 3, pp. 409-412, Mar. 1968.
 """
+
 import math
 
 import numpy as np